+++ src/main/java/org/cactoos/text/AbbreviatedRight.java
@@ -70,7 +70,7 @@ public final class AbbreviatedRight extends TextEnvelope {
             new Flattened(
                 new Ternary<>(
                     new ScalarOf<>(() -> new Sticky(text)),
-                    (Text t) -> new LengthOf(t).value() <= (long) max,
+                    (Text t) -> new LengthOf(t).value() <= max,
                     t -> t,
                     t -> new FormattedText(
                         "%s%s",
